Split command center assets to satisfy static size guards
