+++ internal/web/static/style.css
@@ -5,6 +5,7 @@
 @import url("./chat.css");
 @import url("./ink.css");
 @import url("./overlay.css");
+@import url("./command-center.css");
 @import url("./companion.css");
 @import url("./items.css");
 @import url("./mobile.css");
