+++ internal/web/static_styles_test.go
@@ -17,6 +17,7 @@ func TestStyleCSSImportsSplitStylesheets(t *testing.T) {
 		`@import url("./chat.css");`,
 		`@import url("./ink.css");`,
 		`@import url("./overlay.css");`,
+		`@import url("./command-center.css");`,
 		`@import url("./companion.css");`,
 		`@import url("./items.css");`,
 		`@import url("./mobile.css");`,
@@ -54,6 +55,7 @@ func TestSplitStylesheetsExistAndStayBounded(t *testing.T) {
 		"chat.css",
 		"ink.css",
 		"overlay.css",
+		"command-center.css",
 		"companion.css",
 		"items.css",
 		"mobile.css",
